+++ Backend /src/main/java/com/isa/Backend/dto/AuthenticationRequest.java
@@ -1,6 +1,5 @@
-package com.isa.Backend.controller;
+package com.isa.Backend.dto;
 
-import com.isa.Backend.model.Role;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -10,10 +9,9 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class RegisterRequest {
-    private String firstname;
-    private String lastname;
+public class AuthenticationRequest {
+
     private String email;
     private String password;
-    private Role role;
 }
+
