+++ Backend /src/main/java/com/isa/Backend/dto/RegisterRequest.java
@@ -1,5 +1,6 @@
-package com.isa.Backend.controller;
+package com.isa.Backend.dto;
 
+import com.isa.Backend.model.Role;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -9,9 +10,10 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class AuthenticationRequest {
-
+public class RegisterRequest {
+    private String firstname;
+    private String lastname;
     private String email;
     private String password;
+    private Role role;
 }
-
