+++ Backend /src/main/java/com/isa/Backend/dto/StudentRegRequest.java
@@ -1,4 +1,4 @@
-package com.isa.Backend.controller;
+package com.isa.Backend.dto;
 
 import com.isa.Backend.model.Users;
 import lombok.AllArgsConstructor;
@@ -11,10 +11,12 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class TeacherRegRequest {
-    private String subject;
+public class StudentRegRequest {
+    private String grade;
     private String dateOfBirth;
-    private String mobile;
+    private String address;
     private Users user;
 
 }
+
+
