+++ Backend /src/main/java/com/isa/Backend/exception/ProfileNotFoundException.java
@@ -1,4 +1,4 @@
-package com.isa.Backend.controller;
+package com.isa.Backend.exception;
 
 public class ProfileNotFoundException extends Throwable {
     public ProfileNotFoundException(String s) {
