+++ Backend /src/main/java/com/isa/Backend/model/Users.java
@@ -49,34 +49,6 @@ public class Users implements UserDetails {
         this.password = password;
         this.role = role;
     }
-//
-//    public void setId(Long id) {
-//        this.id = id;
-//    }
-//
-//    public String getFirstName() {
-//        return firstName;
-//    }
-//
-//    public void setFirstName(String firstName) {
-//        this.firstName = firstName;
-//    }
-//
-//    public String getLastName() {
-//        return lastName;
-//    }
-//
-//    public void setLastName(String lastName) {
-//        this.lastName = lastName;
-//    }
-//
-//    public String getEmail() {
-//        return email;
-//    }
-//
-//    public void setEmail(String email) {
-//        this.email = email;
-//    }
 
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
@@ -118,16 +90,4 @@ public class Users implements UserDetails {
     }
 
 
-//    public void setPassword(String password) {
-//        this.password = password;
-//    }
-//
-////
-//    public Role getRole() {
-//        return role;
-//    }
-//
-//    public void setRole(Role role) {
-//        this.role = role;
-//    }
 }
