+++ Backend /src/main/java/com/isa/Backend/repository/ManagementRepository.java
@@ -5,6 +5,5 @@ import com.isa.Backend.model.Management;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ManagementRepository extends JpaRepository<Management, Long> {
-//    Optional<Student> findById(Long id);
 
 }
