+++ Backend /src/main/java/com/isa/Backend/repository/StudentRepository.java
@@ -5,6 +5,5 @@ import com.isa.Backend.model.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StudentRepository extends JpaRepository<Student, Long> {
-//    Optional<Student> findById(Long id);
 
 }
